requirements.txt
django==4.2
motor==3.3.2
pymongo==4.5.0
uvicorn[standard]==0.23.2
httpx==0.25.0
djangorestframework==3.14.0
python-dotenv==1.0.0
